one turn SM training mission
an attempt to copy the IG training mission.
included the new thing: "chronoTrigger: 2". without it marines which exit the vehicle will be lost.
also had to include layer-type sprite of the scout armor, coz there is no normal sprite.
haven't touched the stat gains, so maybe they are too bad or too good.