Voices added
All the armors use the same as their basic equivalent for now (sensor carapace is the same as carapace, jump carapace is the same as jump armor, felinid veteran carapace is the same as basic felinid, etc). May change later if I find better voice lines or better fitting ones that are already there.