Removes the use of Warp Stones in chaos path manufacturing and entry
It was a fun idea, but in testing and through research I've determined that Warp Stones simply don't appear in enough loot tables (only in some ships inherited from base XCOM) to make using them in to gate manufacturing viable.

I would like to revisit some sort of similar resource use for the chaos path in the future, but with a release coming up I'd like to have the player chaos path launch in a state where it's feasible to play as intended.