Updates language files for compatibility with OXCE 8.0 Windows builds
Language entries that utilized in-file line breaks in their entries were causing crashes with the unreleased OXCE 8.0 builds on windows platforms. Commit updates these entries in both the English and Russian language files, though it should be noted that the Russian language file is still extremely out of date otherwise.