Sorted mission scripts.
Mission scripts have been grouped in missionsScripts.rul by type of script (their purpose and what sort of missions they create) to make reading the file easier. Subcategories in the group are not fully sorted, this is prep for further cleanup.

Also research that was dumped at the end of research file has been sorted. It's going to kill me if that keeps happening.